Created subclass NNAgent that is used by pelican and panther NN versions
