+++ .gitignore
@@ -17,6 +17,8 @@ Components/backend/agents/models/*
 Components/plark-game/plark_game/agents/models/*
 Docker/plark_robot_tests/Results/*
 
+Components/agent-training/agent_training/logs
+
 
 
 # Byte-compiled / optimized / DLL files
@@ -153,3 +155,6 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+
+.DS_Store
